Adicionar configuração do tema.
