+++ mobile/src/App.tsx
@@ -1,19 +1,10 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { StatusBar } from 'react-native';
-import { Provider as PaperProvider, DefaultTheme } from 'react-native-paper';
+import { Provider as PaperProvider } from 'react-native-paper';
+
+import theme from './config/theme';
 import { AuthProvider } from './contexts/auth';
-
 import Routes from './routes/routes';
-
-const theme = {
-  ...DefaultTheme,
-  colors: {
-    ...DefaultTheme.colors,
-    primary: '#610094',
-    accent: '#150050',
-    text: 'white',
-  },
-};
 
 const App = () => {
   return (
