+++ mobile/src/config/theme.ts
@@ -1,0 +1,13 @@
+import { DefaultTheme } from 'react-native-paper';
+
+const theme = {
+  ...DefaultTheme,
+  colors: {
+    ...DefaultTheme.colors,
+    primary: '#610094',
+    accent: '#150050',
+    text: 'white',
+  },
+};
+
+export default theme;
